+++ backend/controllers/marketplace_controller.py
@@ -1,15 +1,15 @@
-from backend.controllers.log import create_log
+from backend.controllers.log_controller import write_log
 from backend.dao.marketplace_dao import write, read
-from backend.models.marketplace import Marketplace
+from backend.models.marketplace_model import Marketplace
 
 
 def write_marketplace(name: str, description: str) -> None:
     marketplace = Marketplace(name, description)
     write(marketplace)
-    create_log("Inserted", f"Marketplace - {name}")
+    write_log("Inserted", f"Marketplace - {name}")
 
 
 def read_marketplace() -> list:
     marketplaces = read()
-    create_log('List', 'Marketplace')
+    write_log('List', 'Marketplace')
     return marketplaces
